Add basic Elem class and a weak test
The test doesn't actually aim for the result that we want but at lesat
the machinary is in there to aim for the right thing later.

The main concern at the moment is that it is probably best to go for
separate classes for <self-closing-tags/> and <open></close-tags> so
that we keep them distinct but that makes the api a little uglier.