Support text in Elem model
By trying to_html and falling back to assuming it is text. This stops us
having to wrap all text in a Text() object or something that provides
the to_html method. We could do that but it clutters the API a little
and with little gained as we should only have Elem nodes or unicode.