Handle mixed values & text
Using the 'this is some text and a {value}' syntax like JSX supports.

We also clean up a lot of other parts of the code mostly to make sure
that trailing commas & newlines are added by the parent object rather
than by the children as they are generally context dependent.

It might make sense to move the indentation to be handled by the parent
as often the indentation level is related to whether or not there is a
newline at the end of the object's output. We'll have to see.

We also switch to using the TagChildren compose method which was there
all along but for some reason we weren't actually calling it which is
concerning. This allows us more control when handling the children or at
least moves it to a sensible location.

We also create a Whitespace class so that we can recognise it and skip
it when creating the Elem output because whitespace between tags is
largely irrelevant. As I type this, I realise that it isn't always and
so we're going to have to tackle that next. Eg. the space between two
inline tags in a paragraph is important, though I guess it is always
compressed to one space, maybe that is the important aspect.