Parse all white space as Whitespace
ie. as in object instead of a just a pattern which we throw away.

We then render it all as a single space which is vague similar to what
happens on the html side. Most importantly it preserves the html
behaviour where needed. ie. we don't lose any whitespace between tags
unexpectedly.