Simplify EmptyTag code
As we're not having to match an ending tag of the same name we don't
need to have a custom parse method and can rely on a simple grammar
instead.

We also get rid of the TagName grammar class as it wasn't really serving
a great purpose. Instead we can just use the Symbol class from the
library.