Add simple os.walk based pkd -> py conversion
We want someway of converting packed files over to pure python files. To
match 'jsx' it would make sense to use 'pyx' but that is already taken
by the Cython project, 'pyd' seems to be taken by Windows Python dlls
and so I've gone for pkd. 'pyk' is also an option but 'k' isn't really
the important part of 'packed' for me.

The code walks the tree of the directory provided and creates a match
.py file for every .pkd file. It is set up to execute if you run the
module as main, eg:

    python -m packed .

Will convert all .pkd files below the current directory.

It would be sensible to introduce a watch mode at some point with either
polling or inotify based functionality.