Handle a line with whitespace and no newline
This was a confusing catch bug. I indented the final """ on a multiline
string and everything started breaking because it turns out with had
nothing to handle a string with content and no newline at the end.

This introduces handling for that as a fall back after other options
have been exhausted.

We add simple tests with content & no new line as well as whitespace
and no new line.