Fix setup.py to not reference __version__
We can't reference __version__ in packed.py until we have installed
pypeg2 and we need to read the setup.py to know that pypeg2 needs to be
installed. Not exactly sure what the common solution for this is without
splitting out __version__ into a __init__.py and having the actual
functionality in a submodule which then makes the API a little less
clean. Maybe we put up with the duplication or extract the version from
the .py file but text search rather than import.