Improve attribute handling
It wasn't approaching the attribute parsing with the correct logic for
multiple attributes. We attempt to improve it here. We make sure that
Attributes is capable of handling the case of no attributes (rather than
making the parent use optional(Attributes) to keep the composing code
cleaner and iterating over an empty list rather than testing to see if
Attributes exists at all.