Deduce indentation from input
When finding a PactBlock (which should be renamed!) we have a look at
the indentation at the start of the line and work with that when
creating the output.

We assume that the indentation will be spaces and multiples of 4 we can
make this more forgiving in the future with some kind of Indent class or
object, at the moment the assumption is held through the rest of the
output code too.