Provide 'packed' decorator for Elem -> text conversion
It is nice for people to be able to write:

    def my_func(self):
        return <a attr="value"></a>

But when converted that returns an Elem object instead of the html text.
So rather than disrupt the syntax on the return line, we provide a
decorator so we can do:

    @packed
    def my_func(self):
        return <a attr="value"></a>

And have the correct result.

We make sure to use functools to transfer the meta data from the wrapped
function as systems like Django's template tags use the function name.