Always output tag pairs
I've just tested the previous version which output empty tags as <tag />
where possible but Chrome tripped over the <i attr="value" /> in the
output so I assume my logic is not actually valid. We might need to look
closer at this in the future. It is possible that for html output some
tags must always be one way and others the other way. I'm not sure why I
don't know this.