Don't parse 'packed' syntax in comments
We make sure that we only look for 'packed' syntax to parse if it is on
a line before a '#'. It means we don't process the contents of comments.

It will fail if there is a string with a '#' character in it before the
'packed' syntax in some kind of valid manner.

Our slightly hacky process make me feel that we'd be better with a true
representation of the Python syntax which would be happy to see 'packed'
syntax only at particular points, ie. where a Python value is acceptable
rather than treating the file as near arbitrary text with some 'packed'
syntax in it.