Added docstrings & improved tag & file naming
We change from EmptyTag and NonEmptyTag to SelfClosingTag and PairedTag
which are a little more representative of what we want as NonEmptyTag
really means a tag pair which may or may not have any children or
attributes and so may or maybe be genuinely considered empty.

We also change File to CodeBlock as there is no need for it to represent
an actual file.

Finally we attempt to add reasonable docstrings to all classes &
methods.